Moved circuit setup to UI module
