--- presentation/api/src/commonMain/kotlin/dev/fobo66/valiutchik/presentation/MainPresenter.kt
@@ -1,22 +1,0 @@
-/*
- *    Copyright 2026 Andrey Mukamolov
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package dev.fobo66.valiutchik.presentation
-
-import com.slack.circuit.runtime.presenter.Presenter
-import dev.fobo66.valiutchik.presentation.entity.MainScreen
-
-interface MainPresenter : Presenter<MainScreen.State>
